course-service/requirements.txt
flask
pymongo
pika
python-dotenv
pymongo[srv]
certifi